Fix ci

+++ src/sampler/sampler.cpp
@@ -12,7 +12,7 @@ Sampler* Sampler::Create(Allocator& alloc, const SamplerInfo& si)
         return alloc.new_object<IndependentSampler>(si.spp);
     case SamplerType::stratified:
     {
-        int32 h = std::sqrt(si.spp);
+        int32 h = int32(std::sqrt(si.spp));
         return alloc.new_object<StratifiedSampler>(h, h, true);
     }
